LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/cross-window/open-samedocument-crossorigin-sameorigindomain.sub-expected.txt



FAIL using window.open() to navigate same-document targeting a same-origin-domain (but cross-origin) window assert_equals: navigationType expected "push" but got "replace"
